Workaround for typechecking of Vector interface, allowing compatibility of different type systems.
HexagonalVector's Coord, Distance and Neighbour all implement methods defined upon the Vector interface. The aim is to allow arbitrary coordinate systems to be dropped in later, whether that be 2D orthogonal grids or 5D triangular grids.

Defining these methods upon the interface has the advantage of forcing all Vectors to implement them. However it also requires all vector systems to accept vectors from other systems in order to typecheck. Furthermore as Vector.X is not defined, we can't use those as the field isn't necessarily defined.

As such the easiest workaround is a function to retrieve coordinates. X and Y are being kept exported for now until it's decided how to instantiate a HexagonalVector. It might make more sense to use an array or slice instead of X and Y, provided the compiler will still optimise all the overhead out (questionable).