Reworked system now functions and displays live movements correctly.
The websockets for /worlds/live.json stall when their respective tab is closed. This is fixable but the issue is more that I need to handle things in a Go-alike way. Websocket should be a handled as a goroutine communicating with the rest of the program using channels, not a hacky array of them.